Fixes PHPUnit 11 support
