+++ src/Illuminate/Testing/Constraints/SeeInOrder.php
@@ -4,15 +4,86 @@ namespace Illuminate\Testing\Constraints;
 
 use PHPUnit\Framework\Constraint\Constraint;
 use PHPUnit\Runner\Version;
+use ReflectionClass;
 
-if (str_starts_with(Version::series(), '10')) {
-    class SeeInOrder extends Constraint
+class SeeInOrder extends Constraint
     {
-        use Concerns\SeeInOrder;
+    /**
+     * The string under validation.
+     *
+     * @var string
+     */
+    protected string $content;
+
+    /**
+     * The last value that failed to pass validation.
+     *
+     * @var string
+     */
+    protected string $failedValue;
+
+    /**
+     * Create a new constraint instance.
+     *
+     * @param  string  $content
+     * @return void
+     */
+    public function __construct($content)
+    {
+        $this->content = $content;
     }
-} else {
-    readonly class SeeInOrder extends Constraint
+
+    /**
+     * Determine if the rule passes validation.
+     *
+     * @param  array  $values
+     * @return bool
+     */
+    public function matches($values): bool
     {
-        use Concerns\SeeInOrder;
+        $position = 0;
+
+        foreach ($values as $value) {
+            if (empty($value)) {
+                continue;
+            }
+
+            $valuePosition = mb_strpos($this->content, $value, $position);
+
+            if ($valuePosition === false || $valuePosition < $position) {
+                $this->failedValue = $value;
+
+                return false;
+            }
+
+            $position = $valuePosition + mb_strlen($value);
+        }
+
+        return true;
+    }
+
+    /**
+     * Get the description of the failure.
+     *
+     * @param  array  $values
+     * @return string
+     */
+    public function failureDescription($values): string
+    {
+        return sprintf(
+            'Failed asserting that \'%s\' contains "%s" in specified order.',
+            $this->content,
+            $this->failedValue
+        );
+    }
+
+    /**
+     * Get a string representation of the object.
+     *
+     * @return string
+     */
+    public function toString(): string
+    {
+        return (new ReflectionClass($this))->name;
     }
 }
